projects/data-pipeline-homework/requirements.txt
python-dotenv>=1.0.1
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.2
pyarrow>=17.0
yfinance>=0.2
